PauliOperators

Documentation for PauliOperators.

Base.Matrix

PauliOperators.KetBitString

PauliOperators.KetBitString

PauliOperators.KetBitString

PauliOperators.Pauli

PauliOperators.Pauli

PauliOperators.Pauli

PauliOperators.PauliSum

PauliOperators.PauliSum

Base.:*

Base.:*

Base.:*

Base.:*

Base.:+

Base.:+

Base.:+

Base.:-

Base.:<

Base.:==

Base.:>

Base.Multimedia.display

Base.adjoint

Base.hash

Base.show

Base.show

Base.string

Base.string

Base.sum!

Base.sum!

LinearAlgebra.adjoint!

LinearAlgebra.mul!

PauliOperators.boson_binary_transformation

PauliOperators.commute

PauliOperators.expectation_value_sign

PauliOperators.get_phase

PauliOperators.is_diagonal

PauliOperators.jordan_wigner

PauliOperators.negate

PauliOperators.osum

PauliOperators.osum

PauliOperators.osum

PauliOperators.osum

PauliOperators.otimes

PauliOperators.otimes

PauliOperators.otimes

PauliOperators.otimes

PauliOperators.phasefree

PauliOperators.random_Pauli

PauliOperators.rotate_phase

Base.Matrix — Method

Base.Matrix(p::Pauli{N}) where N

Create dense matrix representation

PauliOperators.KetBitString — Type

An occupation number vector, up to 128 qubits

PauliOperators.KetBitString — Method

KetBitString(N::Integer, v::Integer)

TBW

PauliOperators.KetBitString — Method

KetBitString(vec::Vector{T}) where T<:Union{Bool, Integer}

TBW

PauliOperators.Pauli — Type

In this representation, the Pauli string operator is represented as two binary strings, one for x and one for z.

The format is as follows:

(i)^	heta Z^z1 X^x1 ⊗ Z^z2 X^x2 ⊗ ⋯ ⊗ Z^zN X^xN

Products of operators simply concatonate the left and right strings separately. For example,

XYZIY = 11001|01101

To create a Y operator, bits in the same locations in z and x should be on. This means that we have a phase to keep track of because Z^1*X^1 = iY. As such, we end up working with operators of the form:

(i)^	heta σ1 ⊗ σ2 ⊗ σ3 ⊗ ⋯ ⊗ σN,

where,

σ ∈ {X, iY, Z, I}

PauliOperators.Pauli — Method

Pauli(N::Integer; X=[], Y=[], Z=[])

constructor for creating PauliBoolVec by specifying the qubits where each X, Y, and Z gates exist

PauliOperators.Pauli — Method

Pauli(z::I, x::I) where I<:Integer

TBW

PauliOperators.PauliSum — Type

ops::Dict{Pauli{N},ComplexF64}

A collection of Paulis, joined by addition

PauliOperators.PauliSum — Method

PauliSum(N)

TBW

Base.:* — Method

Base.:*(ps::PauliSum{N}, a::Number) where {N}

TBW

Base.:* — Method

Base.-(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

Multiply two PauliSums.

Base.:* — Method

Base.:*(p::Pauli{N}, c::Number) where {N}

Multiply a Pauli with a number. This returns a PauliSum

Base.:* — Method

Base.:*(p1::Pauli{N}, p2::Pauli{N}) where {N}

Multiply two Pauli's together

Base.:+ — Method

Base.+(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

Add two PauliSums.

Base.:+ — Method

Base.:+(ps::PauliSum{N}, p::Pauli{N}) where {N}

Add a Pauli to a PauliSum.

Base.:+ — Method

Base.:+(p1::Pauli{N}, p2::Pauli{N}) where {N}

Add two Pauli's together. This returns a PauliSum

Base.:- — Method

Base.-(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

Subtract two PauliSums.

Base.:< — Method

Base.:(<)(p1::Pauli{N}, p2::Pauli{N}) where {N}

Check if p1 < p2

Base.:== — Method

Base.:(==)(p1::Pauli{N}, p2::Pauli{N}) where {N}

Check if they are equal, return true or false

Base.:> — Method

Base.:(>)(p1::Pauli{N}, p2::Pauli{N}) where {N}

Check if p1 > p2

Base.Multimedia.display — Method

Base.display(ps::PauliSum)

TBW

Base.adjoint — Method

LinearAlgebra.adjoint(ps::PauliSum{N}) where N

TBW

Base.hash — Method

Base.hash(p::Pauli{N}, h::UInt) where N

Create a hash for a Pauli. Because we want to collect matching operators, with different phases, we don't actually put the phase in the hash

Base.show — Method

Base.show(io::IO, P::Pauli{N}) where N

TBW

Base.show — Method

Base.show(io::IO, P::PauliMask)

TBW

Base.string — Method

Base.string(p::KetBitString{N}) where N

Display, y = iY

Base.string — Method

Base.display(p::Pauli)

Display, y = iY

Base.sum! — Method

Base.sum!(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

Add two PauliSums.

Base.sum! — Method

Base.sum!(p1::PauliSum{N}, p2::Pauli{N}) where {N}

Add a Pauli to a PauliSum.

LinearAlgebra.adjoint! — Method

LinearAlgebra.adjoint!(ps::PauliSum{N}) where N

TBW

LinearAlgebra.mul! — Method

LinearAlgebra.mul!(ps::PauliSum, a::Number)

TBW

PauliOperators.boson_binary_transformation — Method

boson_binary_transformation(nqubits)

TBW

PauliOperators.commute — Method

commute(p1::Pauli{N}, p2::Pauli{N}) where {N}

Check if they commute, return true or false

PauliOperators.expectation_value_sign — Method

expectation_value_sign(p::Pauli{N}, ket::Vector{Bool}) where N

compute expectation value of Pauli o for a product state ket

PauliOperators.get_phase — Method

get_phase(p::Pauli)

Return the phase of the Pauli, i^θ

PauliOperators.is_diagonal — Method

is_diagonal(p::Pauli)

Check if operator is diagonal in the computational (z) basis. E.g., does this operator consist of only I and/or Z?

PauliOperators.jordan_wigner — Method

jordan_wigner(f::Integer, N::Integer)

Transform a single fermionic creation operator to PauliSum.

Arguments

f: orbital index specifying the spin orbital being acted upon by the creation operator, $\hat{a}_f^\dagger$

N: Total number of spin orbitals

PauliOperators.negate — Method

negate(p::Pauli)

Multiply p by -1

PauliOperators.osum — Method

osum(p1::Pauli{N}, p2::Pauli{M}) where {N,M}

Returns the direct sum of two PauliSums

PauliOperators.osum — Method

osum(p1::PauliSum{N}, p2::Pauli{M}) where {N,M}

Returns the direct sum of a PauliSum and a Pauli

PauliOperators.osum — Method

osum(p1::Pauli{N}, p2::PauliSum{M}) where {N,M}

Returns the direct sum of a PauliSum and a Pauli

PauliOperators.osum — Method

osum(p1::Pauli{N}, p2::Pauli{M}) where {N,M}

Returns the direct sum of two Paulis

PauliOperators.otimes — Method

otimes(p1::PauliSum{N}, p2::PauliSum{M}) where {N,M}

TBW

PauliOperators.otimes — Method

otimes(p1::PauliSum{N}, p2::Pauli{M}) where {N,M}

TBW

PauliOperators.otimes — Method

otimes(p::Pauli{N}, ps::PauliSum{M}) where {N,M}

TBW

PauliOperators.otimes — Method

otimes(p1::Pauli{N}, p2::Pauli{M}) where {N,M}

TBW

PauliOperators.phasefree — Method

phasefree(p::Pauli)

TBW

PauliOperators.random_Pauli — Method

random_Pauli(N)

TBW

PauliOperators.rotate_phase — Method

rotate_phase(p::Pauli{N}, θ::Integer) where N

Rotate phase in units of π/2. In otherwords, multiply the phase by i^θ. E.g., mutliplication by -1 is obtained with θ=2.