PauliOperators

Documentation for PauliOperators.

PauliOperators.KetBitString

PauliOperators.KetBitString

PauliOperators.KetBitString

PauliOperators.PauliBitString

PauliOperators.PauliBitString

PauliOperators.PauliBitString

PauliOperators.PauliBitString

Base.:*

Base.:+

Base.:==

Base.show

Base.show

Base.string

Base.string

PauliOperators.commute

PauliOperators.expectation_value_sign

PauliOperators.is_diagonal

PauliOperators.random_PauliBitString

PauliOperators.KetBitString — Type

An occupation number vector, up to 128 qubits

PauliOperators.KetBitString — Method

KetBitString(N::Integer, v::Integer)

TBW

PauliOperators.KetBitString — Method

KetBitString(vec::Vector{T}) where T<:Union{Bool, Integer}

TBW

PauliOperators.PauliBitString — Type

In this representation, the Pauli string operator is represented as two binary strings, one for x and one for z.

The format is as follows: Z^z1 X^x1 ⊗ Z^z2 X^x2 ⊗ ⋯ ⊗ Z^zN X^xN

Products of operators simply concatonate the left and right strings separately. For example,

XYZIY = 11001|01101

PauliOperators.PauliBitString — Method

PauliBitString(N::Integer; X=[], Y=[], Z=[])

constructor for creating PauliBoolVec by specifying the qubits where each X, Y, and Z gates exist

PauliOperators.PauliBitString — Method

PauliBitString(str::String)

TBW

PauliOperators.PauliBitString — Method

PauliBitString(z::I, x::I) where I<:Integer

TBW

Base.:* — Method

Base.:*(p1::PauliBitString{N}, p2::PauliBitString{N}) where {N}

Multiply two PauliBitString's together

Base.:+ — Method

Base.:+(p::PauliBitString{N}, θ::Integer) where N

Add two PauliBitString's together

Base.:== — Method

Base.:(==)(p1::PauliBitString{N}, p2::PauliBitString{N}) where {N}

Check if they are equal, return true or false

Base.show — Method

Base.show(io::IO, P::PauliMask)

TBW

Base.show — Method

Base.show(io::IO, P::PauliBitString{N}) where N

TBW

Base.string — Method

Base.display(p::PauliBitString)

Display, y = iY

Base.string — Method

Base.string(p::KetBitString{N}) where N

Display, y = iY

PauliOperators.commute — Method

commute(p1::PauliBitString{N}, p2::PauliBitString{N}) where {N}

Check if they commute, return true or false

PauliOperators.expectation_value_sign — Method

expectation_value_sign(p::PauliBitString{N}, ket::Vector{Bool}) where N

compute expectation value of PauliBitString o for a product state ket

PauliOperators.is_diagonal — Method

is_diagonal(p::PauliBitString)

Does this operator consist of only I and/or Z?

PauliOperators.random_PauliBitString — Method

random_PauliBitString(N)

TBW