Functions

Base.:* — Method

Base.:*(ps::PauliSum{N}, a::Number) where {N}

TBW

Base.:* — Method

Base.-(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

Multiply two PauliSums.

Base.:* — Method

Base.:*(p::Pauli{N}, KetBitString{N}) where N

TBW

Base.:* — Method

Base.:*(p::Pauli{N}, c::Number) where {N}

Multiply a Pauli with a number. This returns a PauliSum

Base.:* — Method

Base.:*(p1::Pauli{N}, p2::Pauli{N}) where {N}

Multiply two Pauli's together

Base.:* — Method

Base.:*(p::Pauli{N}, in::Array{T}) where {T,N}

TBW

Base.:+ — Method

Base.+(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

Add two PauliSums.

Base.:+ — Method

Base.:+(ps::PauliSum{N}, p::Pauli{N}) where {N}

Add a Pauli to a PauliSum.

Base.:+ — Method

Base.:+(p1::Pauli{N}, p2::Pauli{N}) where {N}

Add two Pauli's together. This returns a PauliSum

Base.:- — Method

Base.-(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

Subtract two PauliSums.

Base.:- — Method

Base.:-(p::ScaledPauli{T,N}) where {T,N}

Negate the p. We could either change the coeff or the pauli. Not sure which is best, negate coeff currently

Base.:< — Method

Base.:(<)(p1::Pauli{N}, p2::Pauli{N}) where {N}

Check if p1 < p2

Base.:== — Method

Base.:(==)(p1::Pauli{N}, p2::Pauli{N}) where {N}

Check if they are equal, return true or false

Base.:> — Method

Base.:(>)(p1::Pauli{N}, p2::Pauli{N}) where {N}

Check if p1 > p2

Base.:≈ — Method

Base.:≈(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

TBW

Base.Multimedia.display — Method

Base.display(ps::PauliSum)

TBW

Base.adjoint — Method

LinearAlgebra.adjoint(ps::PauliSum{N}) where N

TBW

Base.hash — Method

Base.hash(p::Pauli{N}, h::UInt) where N

Create a hash for a Pauli. Because we want to collect matching operators, with different phases, we don't actually put the phase in the hash

Base.show — Method

Base.show(io::IO, P::Pauli{N}) where N

TBW

Base.show — Method

Base.show(io::IO, P::PauliMask)

TBW

Base.string — Method

Base.string(p::KetBitString{N}) where N

Display, y = iY

Base.string — Method

Base.display(p::Pauli)

Display, y = iY

Base.sum! — Method

Base.sum!(p1::PauliSum{N}, p2::PauliSum{N}) where {N}

Add two PauliSums.

Base.sum! — Method

Base.sum!(p1::PauliSum{N}, p2::Pauli{N}) where {N}

Add a Pauli to a PauliSum.

Base.unique! — Method

Base.unique!(spv::Vector{ScaledPauli{T,N}}) where {T,N}

Sort and add duplicates. This modifies spv to be a sorted list of only unique ScaledPauli's

Base.unique — Method

Base.unique(spv::Vector{ScaledPauli{T,N}}) where {T,N}

Sort and add duplicates. This returns a sorted list of only unique ScaledPauli's

LinearAlgebra.adjoint! — Method

LinearAlgebra.adjoint!(ps::PauliSum{N}) where N

TBW

LinearAlgebra.mul! — Method

LinearAlgebra.mul!(ps::PauliSum, a::Number)

TBW

LinearAlgebra.mul! — Method

LinearAlgebra.mul!(C::Matrix{T}, A::Pauli{N}, B::Matrix{T}, α, β) where {T,N}

mul!(C, A, B, α, β) -> C

ABα+Cβ. The result is stored in C by overwriting it. Note that C must not be aliased with either A or B.

LinearAlgebra.mul! — Method

mult!(out::Matrix{T}, p::Pauli{N}, in::Matrix{T}) where {T,N}

Multiply dense vectors in in with p, storing the result in out.

I think this should be much faster if we were to store vectors as rows, so that summing over states acts on contiguous data

PauliOperators.boson_binary_transformation — Method

boson_binary_transformation(nqubits)

TBW

PauliOperators.clip! — Method

clip!(ps::PauliSum; thresh=1e-16)

Delete Pauli's with coeffs smaller than thresh

PauliOperators.commute — Method

commute(p1::Pauli{N}, p2::Pauli{N}) where {N}

Check if they commute, return true or false

PauliOperators.commute — Method

commute(p1::ScaledPauli{T,N}, p2::ScaledPauli{T,N}) where {T,N}

TBW

PauliOperators.expectation_value_sign — Method

expectation_value_sign(p::Pauli{N}, ket::Vector{Bool}) where N

compute expectation value of Pauli o for a product state ket

PauliOperators.get_phase — Method

get_phase(p::Pauli)

Return the phase of the Pauli, i^θ

PauliOperators.is_diagonal — Method

is_diagonal(p::Pauli)

Check if operator is diagonal in the computational (z) basis. E.g., does this operator consist of only I and/or Z?

PauliOperators.jordan_wigner — Method

jordan_wigner(f::Integer, N::Integer)

Transform a single fermionic creation operator to PauliSum.

Arguments

f: orbital index specifying the spin orbital being acted upon by the creation operator, $\hat{a}_f^\dagger$

N: Total number of spin orbitals

PauliOperators.negate — Method

negate(p::Pauli)

Multiply p by -1

PauliOperators.osum — Method

osum(p1::Pauli{N}, p2::Pauli{M}) where {N,M}

Returns the direct sum of two PauliSums

PauliOperators.osum — Method

osum(p1::PauliSum{N}, p2::Pauli{M}) where {N,M}

Returns the direct sum of a PauliSum and a Pauli

PauliOperators.osum — Method

osum(p1::Pauli{N}, p2::PauliSum{M}) where {N,M}

Returns the direct sum of a PauliSum and a Pauli

PauliOperators.osum — Method

osum(p1::Pauli{N}, p2::Pauli{M}) where {N,M}

Returns the direct sum of two Paulis

PauliOperators.otimes — Method

otimes(p1::PauliSum{N}, p2::PauliSum{M}) where {N,M}

TBW

PauliOperators.otimes — Method

otimes(p1::PauliSum{N}, p2::Pauli{M}) where {N,M}

TBW

PauliOperators.otimes — Method

otimes(p::Pauli{N}, ps::PauliSum{M}) where {N,M}

TBW

PauliOperators.otimes — Method

otimes(p1::Pauli{N}, p2::Pauli{M}) where {N,M}

TBW

PauliOperators.phasefree — Method

phasefree(p::Pauli)

TBW

PauliOperators.random_Pauli — Method

random_Pauli(N)

TBW

PauliOperators.rotate_phase — Method

rotate_phase(p::Pauli{N}, θ::Integer) where N

Rotate phase in units of π/2. In otherwords, multiply the phase by i^θ. E.g., mutliplication by -1 is obtained with θ=2.