Types

Base.Matrix — Method

Base.Matrix(spv::Vector{ScaledPauli{T,N}}) where {T,N}

TBW

Base.Matrix — Method

Base.Matrix(ps::PauliSum{N}; T=ComplexF64) where N

Create a dense Matrix of type T

Base.Matrix — Method

Base.Matrix(p::Pauli{N}) where N

Create dense matrix representation

PauliOperators.KetBitString — Type

An occupation number vector, up to 128 qubits

PauliOperators.KetBitString — Method

KetBitString(N::Integer, v::Integer)

TBW

PauliOperators.KetBitString — Method

KetBitString(vec::Vector{T}) where T<:Union{Bool, Integer}

TBW

PauliOperators.Pauli — Type

In this representation, the Pauli string operator is represented as two binary strings, one for x and one for z.

The format is as follows:

(i)^	heta Z^z1 X^x1 ⊗ Z^z2 X^x2 ⊗ ⋯ ⊗ Z^zN X^xN

Products of operators simply concatonate the left and right strings separately. For example,

XYZIY = 11001|01101

To create a Y operator, bits in the same locations in z and x should be on. This means that we have a phase to keep track of because Z^1*X^1 = iY. As such, we end up working with operators of the form:

(i)^	heta σ1 ⊗ σ2 ⊗ σ3 ⊗ ⋯ ⊗ σN,

where,

σ ∈ {X, iY, Z, I}

PauliOperators.Pauli — Method

Pauli(N::Integer; X=[], Y=[], Z=[])

constructor for creating PauliBoolVec by specifying the qubits where each X, Y, and Z gates exist

PauliOperators.Pauli — Method

Pauli(z::I, x::I) where I<:Integer

TBW

PauliOperators.PauliSum — Type

ops::Dict{Pauli{N},ComplexF64}

A collection of Paulis, joined by addition. This uses a Dict to store them, however, the specific use cases should probably dictate the container type, so this will probably be removed.

PauliOperators.PauliSum — Method

PauliSum(N)

TBW

PauliOperators.ScaledPauli — Type

coeff::T
pauli::Pauli{N}

Simply a combination of a Pauli with a coefficient. When sorted, only pauli is considered for the comparisons.